pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teredacta"
version = "0.1.0"
description = "Web interface for Unobfuscator"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "itsdangerous>=2.1.0",
    "bcrypt>=4.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.scripts]
teredacta = "teredacta.__main__:cli"

[tool.setuptools.packages.find]
include = ["teredacta*"]
